requirements.txt
pillow==12.0.0
PySide6==6.10.1